pushed new code of pig game

+++ src/Main.java
@@ -3,7 +3,7 @@
 public class Main {
     public static void main(String[] args) {
         byte a = 30;
-        byte b = (byte) 30*2;
+        byte b = 30*2;
         System.out.println(b);
     }
 }
